% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{covariate_eff_pce}
\alias{covariate_eff_pce}
\title{Average PCE value for all columns in X}
\usage{
covariate_eff_pce(
  W,
  X,
  q,
  x_r = c(-3, 3),
  len = 301,
  d = "sd",
  response = "continuous"
)
}
\arguments{
\item{W}{Weight vector}

\item{X}{Data}

\item{q}{Number of hidden units}

\item{x_r}{x-axis range}

\item{len}{number of breaks for x-axis}

\item{d}{difference value}

\item{response}{Response type: \code{"continuous"} (default) or
\code{"binary"}}
}
\value{
Effect for each input
}
\description{
Average PCE value for all columns in X
}
