% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delta_method}
\alias{delta_method}
\title{Perform delta method for a function FUN to calculate associated uncertainty}
\usage{
delta_method(
  W,
  X,
  y,
  q,
  ind,
  FUN,
  alpha = 0.05,
  x_r = c(-3, 3),
  len = 301,
  ...
)
}
\arguments{
\item{W}{Weight vector}

\item{X}{Data}

\item{y}{Response}

\item{q}{Number of hidden units}

\item{ind}{index of column to plot}

\item{FUN}{function for delta method}

\item{alpha}{significance level}

\item{x_r}{x-axis range}

\item{len}{number of breaks for x-axis}

\item{...}{additional arguments to FUN}
}
\value{
Effect for each input
}
\description{
Perform delta method for a function FUN to calculate associated uncertainty
}
