% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hessian}
\alias{hessian}
\title{Calculate hessian matrix}
\usage{
hessian(W, X, y, q, lambda = 0, response = "continuous")
}
\arguments{
\item{W}{Weight vector}

\item{X}{Data}

\item{y}{Response}

\item{q}{Number of hidden units}

\item{lambda}{Ridge penalty. Default is 0.}

\item{response}{Response type: \code{"continuous"} (default) or
\code{"binary"}}
}
\value{
Hessian matrix
}
\description{
Calculate hessian matrix
}
