% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretnn.R
\name{interpretnn}
\alias{interpretnn}
\alias{interpretnn.default}
\alias{interpretnn.nnet}
\alias{interpretnn.keras.engine.training.Model}
\alias{interpretnn.nn}
\alias{interpretnn.ANN}
\alias{interpretnn.luz_module_fitted}
\title{Statistically-Based Neural Networks}
\usage{
interpretnn(...)

interpretnn.default(object, B = 100, ...)

interpretnn.nnet(object, X, B = 100, ...)

interpretnn.keras.engine.training.Model(object, X, y, B = 100, ...)

interpretnn.nn(object, B = 100, ...)

interpretnn.ANN(object, X, y, B = 100, ...)

interpretnn.luz_module_fitted(object, X, y, B = 100, ...)
}
\arguments{
\item{...}{arguments passed to or from other methods}

\item{object}{nnet object}

\item{B}{number of bootstrap replicates}

\item{X}{matrix of input data}

\item{y}{response variable}
}
\value{
A list with information of the optimal model.
\itemize{
\item \code{interpretnn} - object of class interpretnn.
}

interpretnn object

interpretnn object

interpretnn object

interpretnn object

interpretnn object

interpretnn object

interpretnn object
}
\description{
Return statistically-based outputs for neural networks.
}
