% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis-tests.R
\name{lr_test}
\alias{lr_test}
\title{Likelihood ratio test for inputs}
\usage{
lr_test(X, y, W, q, n_init = 1, unif = 3, maxit = 1000, ...)
}
\arguments{
\item{X}{Data}

\item{y}{Response}

\item{W}{Weight vector}

\item{q}{Number of hidden nodes}

\item{n_init}{Number of initialisations}

\item{unif}{Value for generating random weights}

\item{maxit}{Maximum number of iterations for nnet}

\item{...}{additional arguments to nnet}
}
\value{
Wald hypothesis test for each input
}
\description{
Likelihood ratio test for inputs
}
