% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mlesim}
\alias{mlesim}
\title{Perform m.l.e. simulation for a function FUN to calculate associated uncertainty}
\usage{
mlesim(
  W,
  X,
  y,
  q,
  ind,
  FUN,
  B = 1000,
  alpha = 0.05,
  x_r = c(-3, 3),
  len = 301,
  lambda = 0,
  response = "continuous",
  ...
)
}
\arguments{
\item{W}{Weight vector}

\item{X}{Data}

\item{y}{Response}

\item{q}{Number of hidden units}

\item{ind}{index of column to plot}

\item{FUN}{function for m.l.e. simulation}

\item{B}{number of replicates}

\item{alpha}{significance level}

\item{x_r}{x-axis range}

\item{len}{number of breaks for x-axis}

\item{lambda}{Ridge penalty. Default is 0.}

\item{response}{Response type: \code{"continuous"} (default) or
\code{"binary"}}
}
\value{
Effect for each input
}
\description{
Perform m.l.e. simulation for a function FUN to calculate associated uncertainty
}
