% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_fit.R
\name{nn_fit_torch}
\alias{nn_fit_torch}
\title{Fits various tracks (different random starting values) and chooses best model
using torch}
\usage{
nn_fit_torch(
  x,
  y,
  q,
  n_init,
  inf_crit = "BIC",
  response = "continuous",
  unif = 3,
  maxit = 1000,
  lambda = 0,
  min_delta = 1e-08,
  patience = 10,
  batch_size = 32,
  ...
)
}
\arguments{
\item{x}{Matrix of covariates}

\item{y}{Vector of response}

\item{q}{Number of hidden nodes}

\item{n_init}{Number of random initialisations (tracks)}

\item{inf_crit}{Information criterion: \code{"BIC"} (default), \code{"AIC"} or
\code{"AICc"}}

\item{response}{\code{"continuous"} (default) or \code{"binary"}}

\item{unif}{Random initial values max value}

\item{maxit}{maximum number of iterations for nnet (default = 100)}

\item{lambda}{ridge penalty}

\item{min_delta}{tolerance for early stopping}

\item{patience}{patience for ealy stopping}

\item{batch_size}{batch size}

\item{...}{additional argument for nnet}
}
\value{
The best model from the different tracks
}
\description{
Fits n_init tracks with different initial values and decides on best model
based on information criteria.
}
