% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{nn_grad}
\alias{nn_grad}
\title{Neural network gradient}
\usage{
nn_grad(W, X, q, response = "continuous")
}
\arguments{
\item{W}{Weight vector}

\item{X}{Input vector}

\item{q}{Number of hidden nodes}

\item{response}{Response type: \code{"continuous"} (default) or
\code{"binary"}}
}
\value{
Log-Likelihood value
}
\description{
Neural network gradient
}
