% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{nn_loss}
\alias{nn_loss}
\title{Neural network loss}
\usage{
nn_loss(W, X, y, q, lambda = 0, response = "continuous")
}
\arguments{
\item{W}{Weight vector}

\item{X}{Data}

\item{y}{Response variable}

\item{q}{Number of hidden nodes}

\item{lambda}{Ridge peanlty}

\item{response}{Response type: \code{"continuous"} (default) or
\code{"binary"}}
}
\value{
loss for given neural network
}
\description{
Neural network loss
}
