% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nn_rss}
\alias{nn_rss}
\title{Neural network prediction}
\usage{
nn_rss(W, X, y, q, output = "identity")
}
\arguments{
\item{W}{Weight vector}

\item{X}{Data}

\item{q}{Number of hidden nodes}

\item{output}{Activation function for output unit: \code{"identity"} (default) or
\code{"sigmoid"}}
}
\value{
Prediction for given inputs
}
\description{
Neural network prediction
}
