% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nnet_to_torch}
\alias{nnet_to_torch}
\title{nnet weights to torch weights}
\usage{
nnet_to_torch(nnet_w, p, q)
}
\arguments{
\item{nnet_w}{weights in nnet format}

\item{p}{number of inputs}

\item{q}{number of hidden nodes}
}
\value{
weights in torch format
}
\description{
nnet weights to torch weights
}
