% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pdp_effect}
\alias{pdp_effect}
\title{Partial Dependence Plot for one std. dev. increase}
\usage{
pdp_effect(W, X, q, ind, x_r = c(-3, 3), len = 301)
}
\arguments{
\item{W}{Weight vector}

\item{X}{Data}

\item{q}{Number of hidden units}

\item{ind}{index of column to plot}

\item{x_r}{x-axis range}

\item{len}{number of breaks for x-axis}
}
\value{
Effect for each input
}
\description{
Partial Dependence Plot for one std. dev. increase
}
