% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotci.R
\name{plotci}
\alias{plotci}
\title{Plot Wald Confidence Intervals}
\usage{
plotci(
  object,
  alpha = 0.05,
  which = c(1L:ncol(object$X)),
  colour = 1,
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive(),
  caption = lapply(1:ncol(object$X), function(iter) {
    
    paste0("Confidence Intervals for ", colnames(object$X)[iter])
 }),
  ...
)
}
\arguments{
\item{object}{an object of class \code{statnnet}.}

\item{alpha}{significane level.}

\item{which}{index of plots to be displayed.}

\item{colour}{colour of confidence intervals.}

\item{ask}{ask before displaying each plot.}

\item{caption}{caption for each plot.}

\item{\dots}{arguments to be passed to methods, such as graphical parameters
(see \code{\link{par}}).}
}
\value{
plot of weights and their significance
}
\description{
\code{plotci} It is designed for an inspection of the confidence intervals
for the weights for objects of
class \code{statnn}.
}
