% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotnn.R
\name{plotnn}
\alias{plotnn}
\title{Plot method for neural networks}
\usage{
plotnn(
  x,
  rep = NULL,
  x.entry = NULL,
  x.out = NULL,
  radius = 0.15,
  arrow.length = 0.2,
  intercept = FALSE,
  intercept.factor = 0.4,
  information = TRUE,
  information.pos = 0.1,
  col.entry.synapse = "black",
  col.entry = "black",
  col.hidden = "black",
  col.hidden.synapse = "black",
  col.out = "black",
  col.out.synapse = "black",
  col.intercept = "black",
  col.sig.synapse = "black",
  col.insig.synapse = "lightgrey",
  col.sig.node = "black",
  col.insig.node = "lightgrey",
  fontsize = 12,
  dimension = 6,
  show.weights = FALSE,
  file = NULL,
  rounding = 3,
  alpha = 0.05,
  lambda = 0,
  ...
)
}
\arguments{
\item{x}{an object of class \code{statnnet}}

\item{rep}{repetition of the neural network. If rep="best", the repetition
with the smallest error will be plotted. If not stated all repetitions will
be plotted, each in a separate window.}

\item{x.entry}{x-coordinate of the entry layer. Depends on the arrow.length
in default.}

\item{x.out}{x-coordinate of the output layer.}

\item{radius}{radius of the neurons.}

\item{arrow.length}{length of the entry and out arrows.}

\item{intercept}{a logical value indicating whether to plot the intercept.}

\item{intercept.factor}{x-position factor of the intercept. The closer the
factor is to 0, the closer the intercept is to its left neuron.}

\item{information}{a logical value indicating whether to add the error and
steps to the plot.}

\item{information.pos}{y-position of the information.}

\item{col.entry.synapse}{color of the synapses leading to the input neurons.}

\item{col.entry}{color of the input neurons.}

\item{col.hidden}{color of the neurons in the hidden layer.}

\item{col.hidden.synapse}{color of the weighted synapses.}

\item{col.out}{color of the output neurons.}

\item{col.out.synapse}{color of the synapses leading away from the output
neurons.}

\item{col.intercept}{color of the intercept.}

\item{col.sig.synapse}{color of the significant synapses.}

\item{col.insig.synapse}{color of the insignificant synapses.}

\item{col.sig.node}{color of the significant input nodes.}

\item{col.insig.node}{color of the insignificant input nodes.}

\item{fontsize}{fontsize of the text.}

\item{dimension}{size of the plot in inches.}

\item{show.weights}{a logical value indicating whether to print the
calculated weights above the synapses.}

\item{file}{a character string naming the plot to write to. If not stated,
the plot will not be saved.}

\item{rounding}{number of decimal places to round values.}

\item{alpha}{significane level.}

\item{lambda}{ridge penalty.}

\item{\dots}{arguments to be passed to methods, such as graphical parameters
(see \code{\link{par}}).}
}
\value{
plot of weights and their significance
}
\description{
\code{plotnn} It is designed for an inspection of the weights for objects of
class \code{statnn}.
}
