\name{statnn-methods}
\docType{methods}
\alias{statnn-methods}
\alias{statnn,ANY-method}
\alias{statnn,keras.engine.training.Model-method}
\title{Methods for Function \code{statnn} in Package \pkg{statnn}}
\description{
  Methods for function \code{statnn} in package \pkg{statnn}
}
\section{Methods}{
\describe{

\item{\code{signature(... = "ANY")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(... = "keras.engine.training.Model")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{other possible keyword(s)}
