% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statnn.R
\name{statnn}
\alias{statnn}
\alias{statnn.default}
\alias{statnn.nnet}
\alias{statnn.keras.engine.training.Model}
\alias{statnn.nn}
\alias{statnn.ANN}
\alias{statnn.luz_module_fitted}
\title{Statistically-Based Neural Networks}
\usage{
statnn(...)

statnn.default(object, B = 100, ...)

statnn.nnet(object, X, B = 100, ...)

statnn.keras.engine.training.Model(object, X, y, B = 100, ...)

statnn.nn(object, B = 100, ...)

statnn.ANN(object, X, B = 100, ...)

statnn.luz_module_fitted(object, X, y, B = 100, ...)
}
\arguments{
\item{...}{arguments passed to or from other methods}

\item{object}{nnet object}

\item{B}{number of bootstrap replicates}

\item{X}{matrix of input data}

\item{y}{response variable}
}
\value{
A list with information of the optimal model.
\itemize{
\item \code{statnn} - object of class statnn.
}

statnn object

statnn object

statnn object

statnn object

statnn object

statnn object

statnn object
}
\description{
Return statistically-based outputs for neural networks.
}
