% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{torch_to_nnet}
\alias{torch_to_nnet}
\title{torch weights to nnet weights}
\usage{
torch_to_nnet(torch_w)
}
\arguments{
\item{torch_w}{weights in torch format}
}
\value{
weights in nnet format
}
\description{
torch weights to nnet weights
}
