% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis-tests.R
\name{wald_test}
\alias{wald_test}
\title{Wald test for inputs}
\usage{
wald_test(X, y, W, q)
}
\arguments{
\item{X}{Data}

\item{y}{Response}

\item{W}{Weight vector}

\item{q}{Number of hidden nodes}
}
\value{
Wald hypothesis test for each input
}
\description{
Wald test for inputs
}
