% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{covariate_eff}
\alias{covariate_eff}
\title{Difference in average prediction for values above and below median}
\usage{
covariate_eff(X, W, q)
}
\arguments{
\item{X}{Data}

\item{W}{Weight vector}

\item{q}{Number of hidden units}
}
\value{
Effect for each input
}
\description{
Difference in average prediction for values above and below median
}
