% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_fit.R
\name{nn_fit}
\alias{nn_fit}
\title{Fits various tracks (different random starting values) and chooses best model}
\usage{
nn_fit(
  X,
  y,
  q,
  n_init,
  inf_crit = "BIC",
  lambda = 0,
  response = "continuous",
  unif = 3,
  maxit = 1000,
  pkg = "nnet",
  ...
)
}
\arguments{
\item{X}{Matrix of covariates}

\item{y}{Vector of response}

\item{q}{Number of hidden nodes}

\item{n_init}{Number of random initialisations (tracks)}

\item{inf_crit}{Information criterion: \code{"BIC"} (default), \code{"AIC"} or
\code{"AICc"}}

\item{lambda}{Ridge penalty}

\item{response}{Response type: \code{"continuous"} (default) or
\code{"binary"}}

\item{unif}{Random initial values max value}

\item{maxit}{Maximum number of iterations for nnet (default = 100)}

\item{pkg}{Package for fitting neural network. One of \code{nnet} (default) or
\code{torch}}

\item{...}{additional argument for nnet}
}
\value{
The best model from the different initialisations
}
\description{
Fits n_init tracks with different initial values and decides on best model
based on information criteria.
}
